Move the rematerialize wiring out of update_weight/common.py
These three helpers serve the backuper's sleep/wake lifecycle, not pushing
weights to the rollout engines, and their only dependency was
strip_param_name_prefix from misc_utils - nothing in common.py. Own module
next to actor.py, which also lets the MainCastContext import come up to the
top. Resolves the "move those utils to separate files" TODO.
