+++ miles/backends/megatron_utils/actor.py
@@ -55,8 +55,9 @@ from .initialize import init, is_first_replica_megatron_main_rank
 from .lora_utils import is_lora_enabled, lora_rollout_enabled
 from .model import TrainStepOutcome, forward_only, initialize_model_and_optimizer, save, train
 from .parallel import verify_megatron_parallel_state
+from .rematerialize_utils import build_main_cast_context
 from .replay_utils import register_replay_list_moe
-from .update_weight.common import build_main_cast_context, named_params_and_buffers
+from .update_weight.common import named_params_and_buffers
 from .update_weight.update_weight_from_distributed.broadcast import UpdateWeightFromDistributed
 from .update_weight.update_weight_from_distributed.p2p import UpdateWeightP2P
 from .update_weight.update_weight_from_tensor import UpdateWeightFromTensor
